backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
alembic==1.13.1
pymysql==1.1.1
cryptography==42.0.8
pydantic==2.7.4
pydantic-settings==2.3.4
python-dotenv==1.0.1
httpx==0.27.0
azure-identity==1.17.1
azure-keyvault-secrets==4.8.0
azure-communication-email==1.0.0
applicationinsights==0.11.10
opencensus-ext-azure==1.1.13
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
